+++ parallel-matrix-mult/ptmatmul.c
@@ -2,11 +2,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "macros.h"
+#include <inttypes.h>
 #include "ptmatmul.h"
 
-void add(double** C, u_int ini_cr, u_int ini_cc,
-         double** T, u_int ini_tr, u_int ini_tc,
-         u_int size_r, u_int size_c) {
+void add(double** C, uint64_t ini_cr, uint64_t ini_cc,
+         double** T, uint64_t ini_tr, uint64_t ini_tc,
+         uint64_t size_r, uint64_t size_c) {
     if (!size_r || !size_c)
         return ;
     if (size_r == 1 && size_c == 1) {
@@ -39,30 +40,30 @@ void* matmul_pt_rec(void* arg) {
     double** A = a->A;
     double** B = a->B;
     double** C = a->C;
-    u_int size_ac = a->size_ac;
-    u_int size_ar = a->size_ar;
-    u_int size_bc = a->size_bc;
-    u_int ini_ar = a->ini_ar;
-    u_int ini_ac = a->ini_ac;
-    u_int ini_br = a->ini_br;
-    u_int ini_bc = a->ini_bc;
-    u_int ini_cr = a->ini_cr;
-    u_int ini_cc = a->ini_cc;
-    u_int min_size = a->min_size;
-    //printf("%d, %d, %d, %d, %d\n", size_ac, size_ar, size_bc, ini_ac, ini_bc);
+    uint64_t size_ac = a->size_ac;
+    uint64_t size_ar = a->size_ar;
+    uint64_t size_bc = a->size_bc;
+    uint64_t ini_ar = a->ini_ar;
+    uint64_t ini_ac = a->ini_ac;
+    uint64_t ini_br = a->ini_br;
+    uint64_t ini_bc = a->ini_bc;
+    uint64_t ini_cr = a->ini_cr;
+    uint64_t ini_cc = a->ini_cc;
+    uint64_t min_size = a->min_size;
+    //printf("%lu, %lu, %lu, %lu, %lu\n", size_ac, size_ar, size_bc, ini_ac, ini_bc);
     if (size_ar <= min_size && size_ac <= min_size && size_bc <= min_size) {
-        for (int i = 0; i < size_ar; i++) {
-            for (int k = 0; k < size_ac; k++) {
-                for (int j = 0; j < size_bc; j++)
+        for (uint64_t i = 0; i < size_ar; i++) {
+            for (uint64_t k = 0; k < size_ac; k++) {
+                for (uint64_t j = 0; j < size_bc; j++)
                     C[ini_cr+i][ini_cc+j] += A[ini_ar+i][ini_ac+k]*B[ini_br+k][ini_bc+j];
             }
         }
         free(a);
         return NULL;
     }
-    u_int new_size_ar = size_ar/2;
-    u_int new_size_ac = size_ac/2;
-    u_int new_size_bc = size_bc/2;
+    uint64_t new_size_ar = size_ar/2;
+    uint64_t new_size_ac = size_ac/2;
+    uint64_t new_size_bc = size_bc/2;
     pthread_t t1, t2, t3, t4, t5, t6, t7, t8;
     Argument a_tmp;
 
@@ -125,7 +126,7 @@ void* matmul_pt_rec(void* arg) {
     return NULL;
 }
 
-void matmul_pt(Matrix A, Matrix B, Matrix C, u_int min_size) {
+void matmul_pt(Matrix A, Matrix B, Matrix C, uint64_t min_size) {
     Argument a = create_argument(A->matrix, 0, 0, B->matrix, 0, 0, C->matrix, 0, 0, A->n, A->m, B->m, min_size);
     matmul_pt_rec((void*)a);
 }
